+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.36.0
 * Changed behavior on `result.Err()` on truncated result (returns non-retryable error now, exclude `StreamExecuteScanQuery`)
 * Added `ydb.WithIgnoreTruncated` option for disabling errors on truncated flag
 * Added simple transaction control constructors `table.OnlineReadOnlyTxControl()` and `table.StaleReadOnlyTxControl()`